+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [enhancement](https://github.com/FrankC01/pysui/issues/393) Adds PartyObject encoding. Examples added to ucs_example.py in repo
+- Update ZkSealConfig to support apiKey/apiKeyName and automatic json config migration
 
 ### Fixed